fix delegate/path provider
