+++ app/brave_main_delegate.cc
@@ -8,38 +8,46 @@
 #include <memory>
 #include <optional>
 #include <string>
-#include <unordered_set>
+// #include <unordered_set>
 
-#include "base/base_switches.h"
+// #include "base/base_switches.h"
 #include "base/lazy_instance.h"
 #include "base/path_service.h"
 #include "base/time/time.h"
-#include "brave/app/brave_command_line_helper.h"
 #include "brave/browser/brave_content_browser_client.h"
 #include "brave/common/resource_bundle_helper.h"
 #include "brave/components/brave_component_updater/browser/features.h"
 #include "brave/components/brave_component_updater/browser/switches.h"
+#include "brave/components/brave_sync/buildflags.h"
 #include "brave/components/constants/brave_switches.h"
 #include "brave/components/speedreader/common/buildflags/buildflags.h"
 #include "brave/components/update_client/buildflags.h"
+#include "brave/components/variations/buildflags.h"
 #include "brave/renderer/brave_content_renderer_client.h"
 #include "brave/utility/brave_content_utility_client.h"
 #include "build/build_config.h"
+#include "chrome/app/chrome_main_delegate.h"
 #include "chrome/common/chrome_features.h"
 #include "chrome/common/chrome_paths.h"
 #include "chrome/common/chrome_paths_internal.h"
 #include "chrome/common/chrome_switches.h"
 #include "components/component_updater/component_updater_switches.h"
+#include "components/dom_distiller/core/dom_distiller_switches.h"
 #include "components/embedder_support/switches.h"
 #include "components/sync/base/command_line_switches.h"
 #include "components/variations/variations_switches.h"
-#include "content/public/common/content_switches.h"
+// #include "content/public/common/content_switches.h"
 #include "google_apis/gaia/gaia_switches.h"
 
 #if BUILDFLAG(IS_LINUX)
 #include "base/linux_util.h"
 #endif
 
+#if BUILDFLAG(IS_ANDROID)
+#include "base/android/jni_android.h"
+#include "brave/build/android/jni_headers/BraveQAPreferences_jni.h"
+#include "components/signin/public/base/account_consistency_method.h"
+#endif
 namespace {
 
 const char kBraveOriginTrialsPublicKey[] =
@@ -63,6 +71,29 @@ std::string GetUpdateURLHost() {
   }
   return BUILDFLAG(UPDATER_DEV_ENDPOINT);
 }
+
+#if BUILDFLAG(IS_ANDROID)
+void AdjustSyncServiceUrlForAndroid(std::string* brave_sync_service_url) {
+  DCHECK_NE(brave_sync_service_url, nullptr);
+  const char kProcessTypeSwitchName[] = "type";
+
+  // On Android we can detect data dir only on host process, and we cannot
+  // for example on renderer or gpu-process, because JNI is not initialized
+  // And no sense to override sync service url for them in anyway
+  if (base::CommandLine::ForCurrentProcess()->HasSwitch(
+          kProcessTypeSwitchName)) {
+    // This is something other than browser process
+    return;
+  }
+
+  JNIEnv* env = base::android::AttachCurrentThread();
+  bool b_use_staging_sync_server =
+      Java_BraveQAPreferences_isSyncStagingUsed(env);
+  if (b_use_staging_sync_server) {
+    *brave_sync_service_url = kBraveSyncServiceStagingURL;
+  }
+}
+#endif  // BUILDFLAG(IS_ANDROID)
 
 }  // namespace
 
@@ -113,6 +144,42 @@ content::ContentUtilityClient* BraveMainDelegate::CreateContentUtilityClient() {
 #endif
 }
 
+std::optional<int> BraveMainDelegate::BasicStartupComplete() {
+  auto* command_line = base::CommandLine::ForCurrentProcess();
+  command_line->AppendSwitch(switches::kDisableDomainReliability);
+  command_line->AppendSwitch(switches::kEnableDomDistiller);
+  command_line->AppendSwitch(switches::kEnableDistillabilityService);
+
+  if (!base::CommandLine::ForCurrentProcess()->HasSwitch(
+          embedder_support::kOriginTrialPublicKey)) {
+    command_line->AppendSwitchASCII(embedder_support::kOriginTrialPublicKey,
+                                    kBraveOriginTrialsPublicKey);
+  }
+
+  std::string brave_sync_service_url = BUILDFLAG(BRAVE_SYNC_ENDPOINT);
+#if BUILDFLAG(IS_ANDROID)
+  AdjustSyncServiceUrlForAndroid(&brave_sync_service_url);
+#endif  // BUILDFLAG(IS_ANDROID)
+
+  // Brave's sync protocol does not use the sync service url
+  command_line->AppendSwitchASCII(syncer::kSyncServiceURL,
+                                  brave_sync_service_url.c_str());
+
+  command_line->AppendSwitchASCII(switches::kLsoUrl, kDummyUrl);
+
+  // Brave variations
+  command_line->AppendSwitchASCII(variations::switches::kVariationsServerURL,
+                                  BUILDFLAG(BRAVE_VARIATIONS_SERVER_URL));
+  // Insecure fall-back for variations is set to the same (secure) URL. This is
+  // done so that if VariationsService tries to fall back to insecure url the
+  // check for kHttpScheme in VariationsService::MaybeRetryOverHTTP would
+  // prevent it from doing so as we don't want to use an insecure fall-back.
+  command_line->AppendSwitchASCII(
+      variations::switches::kVariationsInsecureServerURL,
+      BUILDFLAG(BRAVE_VARIATIONS_SERVER_URL));
+
+  return ChromeMainDelegate::BasicStartupComplete();
+}
 void BraveMainDelegate::PreSandboxStartup() {
   ChromeMainDelegate::PreSandboxStartup();
 #if BUILDFLAG(IS_LINUX) || BUILDFLAG(IS_MAC)
@@ -163,20 +230,20 @@ std::optional<int> BraveMainDelegate::PostEarlyInitialization(
     return result;
   }
 
-  BraveCommandLineHelper command_line(base::CommandLine::ForCurrentProcess());
+  auto* command_line = base::CommandLine::ForCurrentProcess();
   std::string update_url = GetUpdateURLHost();
   if (!update_url.empty()) {
-    auto* cmd = base::CommandLine::ForCurrentProcess();
     std::string current_value;
-    if (cmd->HasSwitch(switches::kComponentUpdater)) {
-      current_value = cmd->GetSwitchValueASCII(switches::kComponentUpdater);
-      cmd->RemoveSwitch(switches::kComponentUpdater);
+    if (command_line->HasSwitch(switches::kComponentUpdater)) {
+      current_value =
+          command_line->GetSwitchValueASCII(switches::kComponentUpdater);
+      command_line->RemoveSwitch(switches::kComponentUpdater);
     }
     if (!current_value.empty()) {
       current_value += ',';
     }
 
-    command_line.AppendSwitchASCII(
+    command_line->AppendSwitchASCII(
         switches::kComponentUpdater,
         (current_value + "url-source=" + update_url).c_str());
   }
