+++ components/constants/brave_paths.h
@@ -6,22 +6,17 @@
 #ifndef BRAVE_COMPONENTS_CONSTANTS_BRAVE_PATHS_H_
 #define BRAVE_COMPONENTS_CONSTANTS_BRAVE_PATHS_H_
 
-#include "build/build_config.h"
-
 // This file declares path keys for the brave module.  These can be used
 // with the PathService to access various special directories and files.
 
 namespace brave {
 
 enum {
-  PATH_START = 12000,
+  PATH_START = 20000,
 
   DIR_TEST_DATA,  // Directory where unit test data resides.
   PATH_END
 };
-
-// Call once to register the provider for the path keys defined above.
-void RegisterPathProvider();
 
 }  // namespace brave
